Create FieldCompare
This SPL runs a fieldsummary command to pull all fields from the base search and compare them with field aliases and data model fields. It additionally compares the field alias with data model fields. The purpose is to see is the fields you have correspond to fiels which have aliases added or to field names in data models.